Extract cancellation logic to `ConsultationCancellationService`
- Move consultation cancellation functionality to `ConsultationCancellationService` for better separation of concerns.
- Introduce `IConsultationCancellationService` interface.
- Update `ConsultationController` to use the new service.
- Remove cancellation logic from `ConsultationServiceImpl`.
