+++ src/main/java/com/nexaworks/rafiq/service/consultation/IConsultationCancellationService.java
@@ -1,0 +1,7 @@
+package com.nexaworks.rafiq.service.consultation;
+
+import java.util.UUID;
+
+public interface IConsultationCancellationService {
+    void cancel(UUID id, String reason);
+}
